\name{logistic}
\alias{logistic}
\title{
Logistic sigmoid function
}
\description{
Computes the logistic sigmoid function. Used as a hidden layer transfer function for nonlinear MLP or MONMLP models.
}
\usage{
logistic(x)
}
\arguments{
  \item{x}{
      numeric vector.
  }
}
\seealso{
\code{\link{logistic.prime}}
}

