\name{tansig}
\alias{tansig}
\title{
Hyperbolic tangent sigmoid function
}
\description{
Computes the hyperbolic tangent sigmoid function. Used as a hidden layer transfer function for nonlinear MLP or MONMLP models.
}
\usage{
tansig(x)
}
\arguments{
  \item{x}{
      numeric vector.
  }
}
\seealso{
\code{\link{tansig.prime}}
}
